changed structure for exceptions and apiclient
